requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
docopt==0.6.2
feedgen==0.9.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
packaging==23.0
pluggy==1.0.0
pytest==7.3.0
pytest-watch==4.2.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
watchdog==3.0.0
